refactor: move SpringAnimation to separate file, handle speed on animation level, handle concurrent animations, remove willChange handler (since it doesn't give any benefits), repalce .duration with ONE_FRAME, rename displaylink to link, added TODO list
